Revert "ST-1229"
This reverts commit 8ebd448d8a35d393f34a1f641192531e2fa1b412.

+++ i21Database/dbo/User Defined Types/StagingTransactionLogRadiant.sql
@@ -13,7 +13,7 @@ CREATE TYPE [dbo].[StagingTransactionLogRadiant] AS TABLE(
 	[dtmEndTime] [datetime] NULL,
 	[intEventSequenceID] [int] NULL,
 	[strTrainingModeFlagValue] [nvarchar](40) NULL,
-	[intCashierID] [bigint] NULL,
+	[intCashierID] [int] NULL,
 	[intRegisterID] [int] NULL,
 	[strTillID] [nvarchar](40) NULL,
 	[strOutsideSalesFlagValue] [nvarchar](40) NULL,
